Change Webpack config to support client and server versions
